Split pages down to imports
