+++ static/scripts/modules/Results.js
@@ -1,0 +1,16 @@
+'use strict';
+
+import React from 'react';
+
+
+class Results extends React.Component {
+    render() {
+        return (
+            <div className="results">
+                <h1>Results</h1>
+            </div>
+        );
+    }
+}
+
+export default Results
